build(website): Fix export not found console warnings
* Move Item definition from MenuItem to Menu
